requirements.txt
streamlit>=1.40.0
openai-whisper>=20240930
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
yt-dlp>=2024.04.09
pydub>=0.25.1
pytest>=8.0.0
